requirements.txt
click==8.1.3
colorama==0.4.5
Cython==0.29.32
et-xmlfile==1.1.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
psycopg2-binary==2.9.4
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
xlrd==1.2.0
